requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==3.2.23
djangorestframework==3.15.1
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
